Data.ByteString.Conversion

Documentation

class ToByteString a where

Minimal complete definition

builder

Methods

builder :: a -> Builder

class FromByteString a where

Parse ByteStrings.

Minimal complete definition

parser

Methods

parser :: Parser a

newtype Hex a

Newtype wrapper to parse and produce integral numbers in hexadecimal format

Constructors

Hex

Fields

fromHex :: a

newtype List a

Newtype wrapper to parse and produce a comma separated list of values.

Constructors

List

Fields

fromList :: [a]

toByteString :: ToByteString a => a -> ByteString

toByteString' :: ToByteString a => a -> ByteString

Please note that this needs to convert from a lazy ByteString to a strict one which involves copying the whole string.

runBuilder :: Builder -> ByteString

fromByteString :: FromByteString a => ByteString -> Maybe a

fromByteString' :: FromByteString a => ByteString -> Maybe a

runParser :: Parser a -> ByteString -> Either String a

runParser' :: Parser a -> ByteString -> Either String a