Data.ByteString.Conversion.To

Documentation

class ToByteString a where

Minimal complete definition

builder

Methods

builder :: a -> Builder

toByteString :: ToByteString a => a -> ByteString

toByteString' :: ToByteString a => a -> ByteString

Please note that this needs to convert from a lazy ByteString to a strict one which involves copying the whole string.

runBuilder :: Builder -> ByteString